Teasing out context of PresentedEntity
It is important that the view only interact with one object. By
introducing the collaboration between two objects, business logic is
pushed into the view.

Note, this is not an ideal structure, but is a better approximation.

I still need to tease out a proper Fieldset object with Property/
Predicate objects. But that exact structure is yet to be determined.