Attempting to get Travis to pass
By moving the migrations into the generator, I know that Rails has
been loaded and the context was properly defined.

Prior to this change, when spec/internal did not exist, then the
`rake app:db:create` task was not registered (as it is defined as part
of the spec/internal application).