Adding Property#replace_values and #append_values
These methods are necessary to allow different strategies of work
composition to be leveraged.

At present these strategies may not applied at the correct level, but
it is a start. It also reflects that the model representation of the
data provides a more narrow data management context than the
"traditional" ActiveRecord model.

DLTP-56