Adding form_options to view rendering
Yikes this was a major refactor. At issue was the fact that the
delegator stack was behaving different how I had assumed. I would
attempt to render a form, which would delegate render to the wrapped
work presenter. From that point forward the reference to the object
was the work presenter.

So score one for Python and its method signatures. Today the off hand
comment about decorators by @carljm makes sense.