Partially completing end to end attachment
Attachments are a different critter than other predicates. By the
nature of processing HTML form input the file based predicates are
different.

In the case of other input types (i.e. text, date, integer), those
input fields are passed along the request as plain text. They are a
"simple" data type.

Contrast with the file input type; The data sent is a "complex"
data type. One that requires reading the input buffer, writing it to
another location, and persisting it somewhere that might be very
different from where other predicats can be persisted.