Adding explicit #validate method
In switching from inheritence to composition, I encountered the
a delegator problem.

The point of entry for validation is the Validator#validate method;
this is true of Rails and true of Hydramata::Works::ValidationService.

What is happening prior to this commit is that we have the Validation
object and its collaborating BaseValidator. When the application
attempts to call Validation#validate, the delegator passing along the
behavior to the BaseValidator#validate and the #validate_each message
is received by the BaseValidator instance, and thus uses the initial
behavior.

So in favoring composition, I encountered a problem resolved by
inheritance.