Creating an AttachmentPresenter @WIP
This ties into a configurable behavior that I want to explore; Namely
looking up the appropriate presenter for a given work type, predicate,
and value.

This will also help in the ability to reuse views by presenter type.
The attachment presenter is much more complicated than the other
presenters, but that isn't to say there won't be other classes of
presenters that make sense.

@WIP

This is incomplete (hence the @WIP tag). But it captures an idea.
Note that the Value Presenter and the Attachment Presenter are
duplications of code. I am not satisfied with the inheritance scheme
as I believe there are composable ways of doing this.