Reject empty string values for a property
When a user submits a form via the UI, it is likely that one or more
of the fields will not have input. This could be because it wasn't
required or that the system creates an extra empty field to allow for
additional input.

Given the above, I don't want to be capturing and preserving the non-
present information.

Closes #11